backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
pydantic[email]>=2.6.0,<3.0.0
uvicorn==0.24.0

# Authentication and security
PyJWT>=2.10.1,<3.0.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# HTTP client for Supabase (compatible with existing supabase packages)
httpx[http2]>=0.26,<0.29

# Email services
resend>=0.8.0
aiosmtplib>=3.0.0
jinja2>=3.1.0

# Environment and configuration
python-dotenv==1.0.0
